+++ R/sql.R
@@ -4,8 +4,8 @@
 #' functions to sql functions. Typically a conversion function should escape
 #' all its inputs and return an sql object.
 #'
-#' @param ... Character vectors that will be combined into a single SQL
-#'   expression.
+#' @param ... Character vectors that will be combined into a single SQL vector.
+#' @param x Object to check if it is an sql object.
 #' @export
 sql <- function(...) {
   x <- c_character(...)
@@ -19,31 +19,6 @@ sql <- function(...) {
 #' @export
 `[[.sql` <- function(x, i) {
   sql(NextMethod())
-}
-
-# See setOldClass definition in zzz.R
-
-# c() is also called outside of the dbplyr context so must supply default
-# connection - this seems like a design mistake, and probably an indication
-# that within dbplyr c() should be replaced with a more specific function
-#' @export
-c.sql <- function(..., drop_null = FALSE, con = simulate_dbi()) {
-  input <- list(...)
-
-  if (drop_null) {
-    input <- purrr::compact(input)
-  } # nocov
-
-  out <- unlist(lapply(input, escape, collapse = NULL, con = con))
-  sql(out)
-}
-
-#' @export
-c.ident <- c.sql
-
-#' @export
-unique.sql <- function(x, ...) {
-  sql(NextMethod()) # nocov
 }
 
 #' @rdname sql
@@ -61,16 +36,21 @@ format.sql <- function(x, ...) {
   }
 }
 
-#' @rdname sql
-#' @export
+#' Convert to sql (deprecated)
+#'
+#' `r lifecycle::badge("deprecated")`
+#'
+#' @keywords internal
+#'
 #' @param x Object to coerce
 #' @param con Needed when `x` is directly supplied from the user so that
 #'   schema specifications can be quoted using the correct identifiers.
-as.sql <- function(x, con) UseMethod("as.sql")
-
 #' @export
-as.sql.ident <- function(x, con) x
-#' @export
-as.sql.sql <- function(x, con) x
-#' @export
-as.sql.character <- function(x, con) ident(x)
+as.sql <- function(x, con) {
+  lifecycle::deprecate_warn("2.6.0", "as.sql()", "as_table_path()")
+  if (is_bare_character(x)) {
+    ident(x)
+  } else {
+    x
+  }
+}